05-rag-1/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.13
frozenlist==1.7.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.24
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.45
marshmallow==3.26.1
multidict==6.5.0
mypy_extensions==1.1.0
numpy==2.3.0
openai==1.88.0
orjson==3.10.18
packaging==24.2
portalocker==2.10.1
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypdf==5.6.0
python-dotenv==1.1.0
PyYAML==6.0.2
qdrant-client==1.14.3
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rq==2.4.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
yarl==1.20.1
zstandard==0.23.0
